test: add test cases about statistic modal button events
